+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
