+++ lib/plugins/aws/package/compile/events/api-gateway/lib/cors.js
