+++ lib/plugins/aws/package/compile/events/api-gateway/lib/hack/README.md
