+++ lib/plugins/aws/package/compile/events/api-gateway/lib/hack/updateStage.js
