+++ lib/plugins/aws/package/compile/events/api-gateway/lib/method/authorization.js
