+++ lib/plugins/aws/package/compile/events/api-gateway/lib/method/integration.js
