+++ lib/plugins/aws/package/compile/events/api-gateway/lib/permissions.js
