+++ lib/plugins/aws/package/compile/events/api-gateway/lib/requestValidator.js
