+++ lib/plugins/aws/package/compile/events/api-gateway/lib/stage.js
