+++ lib/plugins/index.js
@@ -33,7 +33,7 @@ module.exports = [
   require('./aws/package/compile/functions.js'),
   require('./aws/package/compile/events/schedule.js'),
   require('./aws/package/compile/events/s3/index.js'),
-  require('./aws/package/compile/events/apiGateway/index.js'),
+  require('./aws/package/compile/events/api-gateway/index.js'),
   require('./aws/package/compile/events/websockets/index.js'),
   require('./aws/package/compile/events/sns.js'),
   require('./aws/package/compile/events/stream.js'),
