+++ test/unit/lib/plugins/aws/package/compile/events/api-gateway/lib/hack/disassociateUsagePlan.test.js
@@ -4,7 +4,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 const Serverless = require('../../../../../../../../../../../lib/serverless');
 const AwsProvider = require('../../../../../../../../../../../lib/plugins/aws/provider');
-const disassociateUsagePlan = require('../../../../../../../../../../../lib/plugins/aws/package/compile/events/apiGateway/lib/hack/disassociateUsagePlan');
+const disassociateUsagePlan = require('../../../../../../../../../../../lib/plugins/aws/package/compile/events/api-gateway/lib/hack/disassociateUsagePlan');
 
 describe('#disassociateUsagePlan()', () => {
   let serverless;
