+++ test/unit/lib/plugins/aws/package/compile/events/api-gateway/lib/method/requestParameters.test.js
