+++ test/unit/lib/plugins/aws/package/compile/events/api-gateway/lib/stage/index.test.js
@@ -5,7 +5,7 @@ const sinon = require('sinon');
 const BbPromise = require('bluebird');
 const _ = require('lodash');
 const childProcess = BbPromise.promisifyAll(require('child_process'));
-const AwsCompileApigEvents = require('../../../../../../../../../../../lib/plugins/aws/package/compile/events/apiGateway');
+const AwsCompileApigEvents = require('../../../../../../../../../../../lib/plugins/aws/package/compile/events/api-gateway');
 const Serverless = require('../../../../../../../../../../../lib/serverless');
 const AwsProvider = require('../../../../../../../../../../../lib/plugins/aws/provider');
 const { createTmpDir } = require('../../../../../../../../../../utils/fs');
